refactors relational db into sql module
