+++ config-test.yml
@@ -1,5 +1,3 @@
-cache:
-  url: "redis://localhost:16379/0"
 db:
   url: "sqlite:///./agent_lab.db"
 vault:
