+++ config.yml
@@ -1,5 +1,3 @@
-cache:
-  url: "redis://localhost:6379/0"
 db:
   url: "postgresql://postgres:postgres@localhost:5432/agent_lab"
 vault:
